// whirlpool/plwhirlpool/Custom-apps/BE/WorldpayConnector/node/typings/configs.d.ts
export interface configs {
  wpUrl: string,
  merchantCode: string,
  installationID: string,
  wpusername: string,
  wppassword: string,
  paymentAppName: string,
  locale: string,
  wpVersion: string,
  gcpappkey: string,
  gcpapptoken: string,
  useXipay: boolean
  xipayUrl: string,
  xipayUsername: string,
  xipayPassword: string,
  xipayMerchantid: string,
  sapmid: string,
  sapgl: string,
  p24DelayToCancel: number
}